requirements.txt
requests>=2.25.1
PyPDF2>=3.0.0
python-docx>=0.8.11
pdfplumber>=0.11.0
Pillow>=9.0.0
fpdf2>=2.7.0
reportlab>=4.0.0 